% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util_redis.R
\name{from_redis_hash}
\alias{from_redis_hash}
\title{Load Redis hash}
\usage{
from_redis_hash(con, name, keys = NULL, f = as.character, missing = NA)
}
\arguments{
\item{con}{Redis connection (\code{redis_api} object)}

\item{name}{Name of the hash}

\item{keys}{Optional list of keys - if omitted we get all}

\item{f}{Function to apply to the \emph{entire} contents of the
hash (not each element)}

\item{missing}{What to replace missing keys with}
}
\description{
Convert contents of a Redis hash into something useful
}

