% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redis.R
\name{hiredis}
\alias{hiredis}
\alias{redis_available}
\title{Interface to Redis}
\usage{
hiredis(host = "127.0.0.1", port = 6379)

redis_available(...)
}
\arguments{
\item{host}{Hostname (default is the localhost)}

\item{port}{Port to connect on (default is Redis' default of 6379)}

\item{...}{Arguments passed from \code{redis_available} to
\code{hiredis}}
}
\description{
Create an interface to Redis, via RcppRedis.  The returned object
includes methods for all Redis commands.
}
\examples{
# Only run if a Redis server is running
if (redis_available()) {
  r <- hiredis()
  r$PING()
  r$SET("foo", "bar")
  r$GET("foo")
}
}

