% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcppredis.R
\name{rcppredis_connection}
\alias{rcppredis_connection}
\title{RcppRedis interface}
\usage{
rcppredis_connection(config = redis_config())
}
\arguments{
\item{config}{A named list of configuration options, as generated
by \code{\link{redis_config}}.  Only \code{host} and \code{port}
are used at present.}
}
\description{
Support for using RcppRedis with RedisAPI.
}
\section{Differences to redux/rrlite}{


 Note that via this interface RcppRedis will not support binary
serialisation (though it does through its own interface) so you
need to use \code{object_to_string}/\code{string_to_object} rather
than \code{object_to_bin}/\code{bin_to_object}.

Treatment of TRUE/FALSE is different as redux will store them as
1/0 but RcppRedis will store as "TRUE"/"FALSE".  Both will be
conveted back to TRUE/FALSE after \code{as.logical} though.

RcppRedis returns strings for Redis statuses (which \emph{are}
just strings) but redux/rrlite add a \code{redis_status}
attribute.
}

