% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdb.R
\name{rdb}
\alias{rdb}
\title{Create an rdb object}
\usage{
rdb(driver, ..., prefix = "rdb:")
}
\arguments{
\item{driver}{A \code{redis_api} object or a function that will
create one when passed arguments \code{...}}

\item{...}{Arguments to the driver}

\item{prefix}{Prefix to add to keys}
}
\description{
Create an rdb object
}
\examples{
\dontrun{
r <- rdb(hiredis)
r$set("foo", runif(10))
r$get("foo")
r$keys()
r$del("foo")
r$keys()
}
}

