% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redis_api.R
\name{redis_api}
\alias{redis_api}
\title{Create a Redis API object}
\usage{
redis_api(x)
}
\arguments{
\item{x}{An object}
}
\description{
Create a Redis API object.  This function is designed to be used
from other packages, and not designed to be used directly by
users.
}

