% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/objects.R
\name{redis_object}
\alias{redis_object}
\alias{redis_object_del}
\alias{redis_object_exists}
\alias{redis_object_get}
\alias{redis_object_set}
\alias{redis_object_type}
\title{Read/write R objects to Redis}
\usage{
redis_object_set(key, object, db, ...)

redis_object_get(key, db, ...)

redis_object_del(key, db, ...)

redis_object_exists(key, db, ...)

redis_object_type(key, db, ...)
}
\arguments{
\item{key}{The key to store it in}

\item{object}{An R object (must be a data.frame at present)}

\item{db}{The database to store it in}

\item{...}{Additional arguments to methods}
}
\description{
Support for writing and reading R objects (currently data.frames)
to a redis instance.
}

