% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redis.R
\name{hiredis}
\alias{hiredis}
\alias{redis_available}
\title{Interface to Redis}
\usage{
hiredis(...)

redis_available(...)
}
\arguments{
\item{...}{Named configuration options passed to
  \code{\link{redis_config}}, used to create the environment
  (notable keys include \code{host}, \code{port}, and the
  environment variable \code{REDIS_URL})}

\item{...}{Arguments passed from \code{redis_available} to
\code{hiredis}}
}
\description{
Create an interface to Redis, via redux.  The returned object
includes methods for all Redis commands.
}
\examples{
# Only run if a Redis server is running
if (redis_available()) {
  r <- hiredis()
  r$PING()
  r$SET("foo", "bar")
  r$GET("foo")
}
}

