% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redis.R
\name{hiredis}
\alias{hiredis}
\alias{redis_available}
\title{Interface to Redis}
\usage{
hiredis(host = Sys.getenv("REDIS_HOST", "127.0.0.1"),
  port = Sys.getenv("REDIS_PORT", 6379))

redis_available(...)
}
\arguments{
\item{host}{Hostname.  Default is to look up the value of the
environment variable \code{REDIS_HOST} and then to default to the
value \code{127.0.0.1} (localhost).}

\item{port}{Port to connect on.  Default is to look up the value
of the environment variable \code{REDIS_PORT} and then to default
to the value "6379".}

\item{...}{Arguments passed from \code{redis_available} to
\code{hiredis}}
}
\description{
Create an interface to Redis, via RcppRedis.  The returned object
includes methods for all Redis commands.
}
\examples{
# Only run if a Redis server is running
if (redis_available()) {
  r <- hiredis()
  r$PING()
  r$SET("foo", "bar")
  r$GET("foo")
}
}

