% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcppredis.R
\name{rcppredis_connection}
\alias{rcppredis_connection}
\alias{rcppredis_hiredis}
\title{RcppRedis interface}
\usage{
rcppredis_connection(config = redis_config())

rcppredis_hiredis(..., version = NULL)
}
\arguments{
\item{config}{A named list of configuration options, as generated
by \code{\link{redis_config}}.  Only \code{host} and \code{port}
are used at present.}

\item{...}{arguments passed through to \code{\link{redis_config}}.
Can include a named argument \code{config} or named arguments
\code{host}, \code{port}.}

\item{version}{Version of the RedisAPI to generate.  If given as a
numeric version (or something that can be coerced into one.  If
given as \code{TRUE}, then we query the Redis server for its
version and generate only commands supported by the server.}
}
\description{
Support for using RcppRedis with RedisAPI.
}
\section{Differences to redux/rrlite}{


 Note that via this interface RcppRedis will not support binary
serialisation (though it does through its own interface) so you
need to use \code{object_to_string}/\code{string_to_object} rather
than \code{object_to_bin}/\code{bin_to_object}.

Treatment of TRUE/FALSE is different as redux will store them as
1/0 but RcppRedis will store as "TRUE"/"FALSE".  Both will be
conveted back to TRUE/FALSE after \code{as.logical} though.

RcppRedis returns strings for Redis statuses (which \emph{are}
just strings) but redux/rrlite add a \code{redis_status}
attribute.
}

