% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redis.R
\name{redis_api}
\alias{redis_api}
\title{Create a Redis API object}
\usage{
redis_api(x, host, port, type)
}
\arguments{
\item{x}{An object}

\item{host}{String indicating the hostnae}

\item{port}{String indicating the port}

\item{type}{String indicating the type}
}
\description{
Create a Redis API object
}
\details{
This function is designed to be used from other packages, and not
designed to be used directly by users.  \code{host}, \code{port}
and \code{type} are never used except for being stored in the
returned object.
}

