% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/redis.R
\name{redis_api}
\alias{redis_api}
\title{Create a Redis API object}
\usage{
redis_api(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed through to methods}
}
\description{
Create a Redis API object
}

