% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scan.R
\name{scan_apply}
\alias{scan_apply}
\alias{scan_del}
\alias{scan_find}
\title{Iterate over keys using SCAN}
\usage{
scan_apply(con, callback, pattern = NULL, count = NULL)

scan_del(con, pattern, count = NULL)

scan_find(con, pattern, count = NULL)
}
\arguments{
\item{con}{A \code{redis_api} object}

\item{callback}{Function that takes a character vector of keys and
does something useful to it.  \code{con$DEL} is one option here to
delete keys that match a pattern.}

\item{pattern}{Optional pattern to use.}

\item{count}{Optional step size (default is Redis' default which
is 10)}
}
\description{
Support for iterating with \code{SCAN}.  Note that this will
generalise soon to support collecting output, \code{SSCAN} and
other variants, etc.
}
\details{
The functions \code{scan_del} and \code{scan_find} are example
functions that delete and find all keys corresponding to a given
pattern.
}

